phase2/ChatGPT/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voronoi-tk-app"
version = "0.1.0"
description = "Voronoi diagram generator with Tkinter GUI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
dev = ["pytest>=8.0"]
image = ["Pillow>=10.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
